+++ Lib/Generic/adf_nativ.c
@@ -20,7 +20,6 @@
  *  Foundation, Inc., 51 Franklin St, Fifth Floor, Boston, MA  02110-1301  USA
  *
  */
- */
 
 #include<stdio.h>
 #include<stdlib.h>
@@ -65,7 +64,7 @@ RETCODE myInitDevice(struct Device* dev, char* name,BOOL ro)
  * myReadSector
  *
  */
-RETCODE myReadSector(struct Device *dev, long n, int size, unsigned char* buf)
+RETCODE myReadSector(struct Device *dev, int32_t n, int size, unsigned char* buf)
 {
      return RC_OK;   
 }
@@ -75,7 +74,7 @@ RETCODE myReadSector(struct Device *dev, long n, int size, unsigned char* buf)
  * myWriteSector
  *
  */
-RETCODE myWriteSector(struct Device *dev, long n, int size, unsigned char* buf)
+RETCODE myWriteSector(struct Device *dev, int32_t n, int size, unsigned char* buf)
 {
     return RC_OK;
 }
