+++ Lib/adf_disk.c
@@ -42,7 +42,7 @@
 
 extern struct Env adfEnv;
 
-unsigned long bitMask[32] = { 
+uint32_t bitMask[32] = { 
     0x1, 0x2, 0x4, 0x8,
 	0x10, 0x20, 0x40, 0x80,
     0x100, 0x200, 0x400, 0x800,
@@ -132,7 +132,7 @@ void adfVolumeInfo(struct Volume *vol)
 		printf ("DIRCACHE ");
 	putchar('\n');
 
-    printf("Free blocks = %ld\n", adfCountFreeBlocks(vol));
+    printf("Free blocks = %d\n", adfCountFreeBlocks(vol));
     if (vol->readOnly)
         printf("Read only\n");
     else
@@ -140,13 +140,13 @@ void adfVolumeInfo(struct Volume *vol)
  	
     /* created */
 	adfDays2Date(root.coDays, &year, &month, &days);
-    printf ("created %d/%02d/%02d %ld:%02ld:%02ld\n",days,month,year,
+    printf ("created %d/%02d/%02d %d:%02d:%02d\n",days,month,year,
 	    root.coMins/60,root.coMins%60,root.coTicks/50);	
 	adfDays2Date(root.days, &year, &month, &days);
-    printf ("last access %d/%02d/%02d %ld:%02ld:%02ld,   ",days,month,year,
+    printf ("last access %d/%02d/%02d %d:%02d:%02d,   ",days,month,year,
 	    root.mins/60,root.mins%60,root.ticks/50);	
 	adfDays2Date(root.cDays, &year, &month, &days);
-    printf ("%d/%02d/%02d %ld:%02ld:%02ld\n",days,month,year,
+    printf ("%d/%02d/%02d %d:%02d:%02d\n",days,month,year,
 	    root.cMins/60,root.cMins%60,root.cTicks/50);	
 }
 
@@ -159,7 +159,7 @@ void adfVolumeInfo(struct Volume *vol)
  */
 struct Volume* adfMount( struct Device *dev, int nPart, BOOL readOnly )
 {
-    long nBlock;
+    int32_t nBlock;
     struct bRootBlock root;
 	struct bBootBlock boot;
 	struct Volume* vol;
@@ -173,7 +173,7 @@ struct Volume* adfMount( struct Device *dev, int nPart, BOOL readOnly )
 	vol->dev = dev;
     vol->mounted = TRUE;
 
-/*printf("first=%ld last=%ld root=%ld\n",vol->firstBlock,
+/*printf("first=%d last=%d root=%d\n",vol->firstBlock,
  vol->lastBlock, vol->rootBlock);
 */
     if (adfReadBootBlock(vol, &boot)!=RC_OK) {
@@ -235,7 +235,7 @@ void adfUnMount(struct Volume *vol)
  *
  * 
  */
-struct Volume* adfCreateVol( struct Device* dev, long start, long len, 
+struct Volume* adfCreateVol( struct Device* dev, int32_t start, int32_t len, 
     char* volName, int volType )
 {
     struct bBootBlock boot;
@@ -258,7 +258,7 @@ struct Volume* adfCreateVol( struct Device* dev, long start, long len,
     vol->firstBlock = (dev->heads * dev->sectors)*start;
     vol->lastBlock = (vol->firstBlock + (dev->heads * dev->sectors)*len)-1;
     vol->rootBlock = (vol->lastBlock - vol->firstBlock+1)/2;
-/*printf("first=%ld last=%ld root=%ld\n",vol->firstBlock,
+/*printf("first=%d last=%d root=%d\n",vol->firstBlock,
  vol->lastBlock, vol->rootBlock);
 */
     vol->curDirPtr = vol->rootBlock;
@@ -347,7 +347,7 @@ printf("%3d %x, ",i,vol->bitmapTable[0]->map[i]);
 
     if (adfEnv.useProgressBar)
         (*adfEnv.progressBar)(100);
-/*printf("free blocks %ld\n",adfCountFreeBlocks(vol));*/
+/*printf("free blocks %d\n",adfCountFreeBlocks(vol));*/
 
     /* will be managed by adfMount() later */
     adfFreeBitmap(vol);
@@ -366,10 +366,10 @@ printf("%3d %x, ",i,vol->bitmapTable[0]->map[i]);
  * read logical block
  */
 RETCODE
-adfReadBlock(struct Volume* vol, long nSect, unsigned char* buf)
+adfReadBlock(struct Volume* vol, int32_t nSect, unsigned char* buf)
 {
   /*    char strBuf[80];*/
-    long pSect;
+    int32_t pSect;
     struct nativeFunctions *nFct;
     RETCODE rc;
 
@@ -384,21 +384,21 @@ adfReadBlock(struct Volume* vol, long nSect, unsigned char* buf)
     if (adfEnv.useRWAccess)
         (*adfEnv.rwhAccess)(pSect,nSect,FALSE);
 
-/*printf("psect=%ld nsect=%ld\n",pSect,nSect);*/
-/*    sprintf(strBuf,"ReadBlock : accessing logical block #%ld", nSect);	
+/*printf("psect=%d nsect=%d\n",pSect,nSect);*/
+/*    sprintf(strBuf,"ReadBlock : accessing logical block #%d", nSect);	
     (*adfEnv.vFct)(strBuf);
 */
     if (pSect<vol->firstBlock || pSect>vol->lastBlock) {
         (*adfEnv.wFct)("adfReadBlock : nSect out of range");
         
     }
-/*printf("pSect R =%ld\n",pSect);*/
+/*printf("pSect R =%d\n",pSect);*/
     nFct = adfEnv.nativeFct;
     if (vol->dev->isNativeDev)
         rc = (*nFct->adfNativeReadSector)(vol->dev, pSect, 512, buf);
     else
         rc = adfReadDumpSector(vol->dev, pSect, 512, buf);
-/*printf("rc=%ld\n",rc);*/
+/*printf("rc=%d\n",rc);*/
     if (rc!=RC_OK)
         return RC_ERROR;
     else
@@ -410,9 +410,9 @@ adfReadBlock(struct Volume* vol, long nSect, unsigned char* buf)
  * adfWriteBlock
  *
  */
-RETCODE adfWriteBlock(struct Volume* vol, long nSect, unsigned char *buf)
+RETCODE adfWriteBlock(struct Volume* vol, int32_t nSect, unsigned char *buf)
 {
-    long pSect;
+    int32_t pSect;
     struct nativeFunctions *nFct;
     RETCODE rc;
 
@@ -427,7 +427,7 @@ RETCODE adfWriteBlock(struct Volume* vol, long nSect, unsigned char *buf)
     }
 
     pSect = nSect+vol->firstBlock;
-/*printf("write nsect=%ld psect=%ld\n",nSect,pSect);*/
+/*printf("write nsect=%d psect=%d\n",nSect,pSect);*/
 
     if (adfEnv.useRWAccess)
         (*adfEnv.rwhAccess)(pSect,nSect,TRUE);
