+++ Lib/adf_link.h
@@ -27,9 +27,10 @@
  */
 
 #include"prefix.h"
+#include <stdint.h>
 
 PREFIX RETCODE adfBlockPtr2EntryName(struct Volume *vol, SECTNUM nSect, SECTNUM lPar, 
-	char **name, long *size);
+	char **name, int32_t *size);
 
 
 #endif /* ADF_LINK_H */
